requirements/requirements.txt
ovos-plugin-manager>=0.0.1a2
googletrans-py
